+++ R/dbAppendStream.R
@@ -1,0 +1,7 @@
+#' @include dbAppendTable.R
+#' @name dbAppendTable
+#' @aliases dbAppendStream
+#' @export
+setGeneric("dbAppendStream",
+  def = function(conn, name, value, ...) standardGeneric("dbAppendStream")
+)