+++ R/dbClearResult_DBIResultStream.R
@@ -1,0 +1,8 @@
+#' @rdname hidden_aliases
+#' @usage NULL
+dbClearResult_DBIResultStream <- function(res, ...) {
+  dbClearResult(res@result, ...)
+}
+#' @rdname hidden_aliases
+#' @export
+setMethod("dbClearResult", signature("DBIResultStream"), dbClearResult_DBIResultStream)