+++ R/dbSendQueryStream.R
@@ -4,16 +4,16 @@
 #' con <- dbConnect(RSQLite::SQLite(), ":memory:")
 #'
 #' dbWriteTable(con, "mtcars", mtcars)
-#' rs <- dbSendQueryArrow(con, "SELECT * FROM mtcars WHERE cyl = 4")
+#' rs <- dbSendQueryStream(con, "SELECT * FROM mtcars WHERE cyl = 4")
 #' dbFetch(rs)
 #' dbClearResult(rs)
 #'
 #' dbDisconnect(con)
 #' @export
-setGeneric("dbSendQueryArrow",
+setGeneric("dbSendQueryStream",
   def = function(conn, statement, ...) {
     require_arrow()
-    standardGeneric("dbSendQueryArrow")
+    standardGeneric("dbSendQueryStream")
   },
-  valueClass = "DBIResultArrow"
+  valueClass = "DBIResultStream"
 )